requirements.txt
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.3.3
apache-airflow-providers-apache-spark==3.0.0
apache-airflow-providers-common-sql==1.0.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.1.0
apispec==3.3.2
argcomplete==2.0.0
attrs==20.3.0
Babel==2.10.3
blinker==1.5
cachelib==0.9.0
cattrs==1.10.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.0
cron-descriptor==1.2.31
croniter==1.3.5
cryptography==37.0.4
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
Flask==2.1.2
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==2.0.0
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
packaging==21.3
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.1
py4j==0.10.9.5
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.0
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==2.3.3
